python_code/requirements.txt
MouseInfo==0.1.2
Pillow==6.2.1
PyAutoGUI==0.9.48
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyperclip==1.7.0
PyQt5==5.14.0
PyQt5-sip==12.7.0
PyRect==0.1.4
PyScreeze==0.1.25
pyserial==3.4
PyTweening==1.0.3
pyinstaller==4.3
pyqtgraph==0.12.1
scipy==1.6.3
pandas==1.2.5
numpy==1.21.0